Created module "networkPopen_weather_api" to separate it from "network_caller" as they don't rely on each other.
