+++ network_caller/build.gradle.kts
@@ -1,18 +1,12 @@
 import com.zywczas.buildutils.ModulesUtils
 import com.zywczas.buildutils.Versions
-import com.zywczas.buildutils.getEnvElseLocal
 
 plugins {
     alias(libs.plugins.kotlin.multiplatform)
     alias(libs.plugins.android.kotlin.multiplatform.library)
     alias(libs.plugins.compose.multiplatform)
-    alias(libs.plugins.gmazzo)
     alias(libs.plugins.kotlin.compose.compiler)
     alias(libs.plugins.detekt)
-}
-
-buildConfig {
-    buildConfigField("OPENWEATHERMAP_API_KEY", getEnvElseLocal("openweathermapApiKey", project))
 }
 
 kotlin {
