--- network_caller/src/commonMain/kotlin/com/zywczas/networkcaller/ApiKeys.kt
@@ -1,7 +1,0 @@
-package com.zywczas.networkcaller
-
-import Weather.network_caller.BuildConfig
-
-object ApiKeys {
-    const val OPEN_WEATHER_MAP = BuildConfig.OPENWEATHERMAP_API_KEY
-}
