+++ network_open_weather_api/src/commonMain/kotlin/com/zywczas/networkopenweatherapi/WeatherApiConstants.kt
@@ -1,6 +1,5 @@
-package com.zywczas.networkcaller.openweatherapi
+package com.zywczas.networkopenweatherapi
 
-//todo move weather api values to separate module
 object WeatherApiConstants {
     const val WEATHER_API_BASE_URL = "api.openweathermap.org/"
     const val LOCATIONS_RESULTS_LIMIT = 5
