+++ network_open_weather_api/src/commonMain/kotlin/com/zywczas/networkopenweatherapi/WeatherApiKeys.kt
@@ -1,0 +1,7 @@
+package com.zywczas.networkopenweatherapi
+
+import Weather.network_open_weather_api.BuildConfig
+
+object WeatherApiKeys {
+    const val OPEN_WEATHER_MAP = BuildConfig.OPENWEATHERMAP_API_KEY
+}
