+++ network_open_weather_api/src/commonMain/kotlin/com/zywczas/networkopenweatherapi/WeatherApiLanguage.kt
@@ -1,4 +1,4 @@
-package com.zywczas.networkcaller.openweatherapi
+package com.zywczas.networkopenweatherapi
 
 enum class WeatherApiLanguage(val value: String) {
     English("en"),
