+++ settings.gradle.kts
@@ -29,3 +29,4 @@ include(":network_places")
 include(":store_history")
 include(":umbrella")
 include(":common_utils")
+include(":network_open_weather_api")
